Update ProfileCard to be stacked on each other
The ProfileCards are viewed one after another by
vertically scrolling through them. The like and
pass buttons would be rendered useless in such a
case.

Let's stack the cards on top on each other and
depending on which button is clicked, display
the card which corresponds to the current index
to be displayed.